Crate substrate_service

Substrate service. Starts a thread that spins up the network, client, and extrinsic pool. Manages communication between them.

Re-exports

pub use self::error::Error;

pub use config::Configuration;

pub use transaction_pool::txpool;

Modules

chain_ops
Chain utilities.

config
Service configuration.

error
Errors that can occur during the service operation.

Macros

export_blocks
Defines the logic for an operation exporting blocks within a range. Export blocks

import_blocks
Defines the logic for an operation importing blocks from some known import. Import blocks

revert_chain
Revert the chain some number of blocks.

Structs

ChainSpec
A configuration of a chain. Can be used to build a genesis block.

NetworkStatus
Overview status of the network.

NewService
Substrate service.

Roles
Bitmask of the roles that a node fulfills.

RpcMetadata
RPC Metadata.

RpcSession
An RPC session. Used to perform in-memory RPC queries (ie. RPC queries that don't go through the HTTP or WebSockets server).

ServiceBuilder
Aggregator for the components required to build a service.

SpawnTaskHandle
An handle for spawning tasks in the service.

TransactionPool
Extrinsics pool that performs validation.

TransactionPoolAdapter
Transaction pool adapter.

TransactionPoolOptions
Pool configuration options.

Enums

PruningMode
Pruning mode.

Traits

AbstractService
Abstraction over a Substrate service.

ChainApi
Concrete extrinsic validation and query logic.

ChainSpecExtension
A collection of ChainSpec extensions.

IntoPoolError
Transaction pool error conversion.

RuntimeGenesis
A set of traits for the runtime genesis config.

ServiceBuilderExport
Implemented on ServiceBuilder. Allows exporting blocks once you have given all the required components to the builder.

ServiceBuilderImport
Implemented on ServiceBuilder. Allows importing blocks once you have given all the required components to the builder.

ServiceBuilderRevert
Implemented on ServiceBuilder. Allows reverting the chain once you have given all the required components to the builder.

Type Definitions

FinalityNotifications
A stream of block finality notifications.

Properties
Arbitrary properties defined in chain spec as a JSON object

TaskExecutor
Alias for a an implementation of futures::future::Executor.