Struct substrate_service::Service

pub struct Service<Components: Components> {
    pub config: FactoryFullConfiguration<Components::Factory>,
    // some fields omitted
}

Substrate service.

Fields
config: FactoryFullConfiguration<Components::Factory>

Configuration of this Service

Methods

impl<Components: Components> Service<Components>

pub fn new(
config: FactoryFullConfiguration<Components::Factory>,
task_executor: TaskExecutor
) -> Result<Self, Error>

Creates a new service.

pub fn authority_key(&self) -> Option<Pair>

give the authority key, if we are an authority and have a key

pub fn telemetry(&self) -> Option<Arc<Telemetry>>

return a shared instance of Telemtry (if enabled)

impl<Components> Service<Components> where
Components: Components,

pub fn client(&self) -> Arc<ComponentClient<Components>>

Get shared client instance.

pub fn network(&self) -> Arc<NetworkService<Components::Factory>>

Get shared network instance.

pub fn transaction_pool(
&self
) -> Arc<TransactionPool<Components::TransactionPoolApi>>

Get shared transaction pool instance.

pub fn inherents_pool(
&self
) -> Arc<InherentsPool<ComponentExtrinsic<Components>>>

Get shared inherents pool instance.

pub fn keystore(&self) -> &Keystore

Get shared keystore.

pub fn on_exit(&self) -> Exit

Get a handle to a future that will resolve on exit.