Struct substrate_service::ServiceBuilder

pub struct ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend> { /* fields omitted */ }

Aggregator for the components required to build a service.

Usage

Call ServiceBuilder::new_full or ServiceBuilder::new_light, then call the various with_ methods to add the required components that you built yourself:

with_select_chain

with_import_queue

with_network_protocol

with_finality_proof_provider

with_transaction_pool

After this is done, call build to construct the service.

The order in which the with_* methods are called doesn't matter, as the correct binding of generics is done when you call build.

Methods

impl<TCfg, TGen, TCSExt> ServiceBuilder<(), (), TCfg, TGen, TCSExt, (), (), (), (), (), (), (), (), (), (), ()> where
TGen: RuntimeGenesis,
TCSExt: Extension,

pub fn new_full<TBl: BlockT<Hash = H256>, TRtApi, TExecDisp: NativeExecutionDispatch>(
config: Configuration<TCfg, TGen, TCSExt>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, Client<Backend<TBl>, LocalCallExecutor<Backend<TBl>, NativeExecutor<TExecDisp>>, TBl, TRtApi>, Arc<OnDemand<TBl>>, (), (), BoxFinalityProofRequestBuilder<TBl>, (), (), (), (), FullRpcBuilder<TBl, TRtApi, TExecDisp>, Backend<TBl>>, Error>

Start the service builder with a configuration.

pub fn new_light<TBl: BlockT<Hash = H256>, TRtApi, TExecDisp: NativeExecutionDispatch + 'static>(
config: Configuration<TCfg, TGen, TCSExt>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, Client<Backend<LightStorage<TBl>, Blake2Hasher>, GenesisCallExecutor<Backend<LightStorage<TBl>, Blake2Hasher>, LocalCallExecutor<Backend<LightStorage<TBl>, Blake2Hasher>, NativeExecutor<TExecDisp>>>, TBl, TRtApi>, Arc<OnDemand<TBl>>, (), (), BoxFinalityProofRequestBuilder<TBl>, (), (), (), (), LightRpcBuilder<TBl, TRtApi, TExecDisp>, Backend<LightStorage<TBl>, Blake2Hasher>>, Error>

Start the service builder with a configuration.

impl<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend> ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>

pub fn client(&self) -> &Arc<TCl>

Returns a reference to the client that was stored in this builder.

pub fn backend(&self) -> &Arc<Backend>

Returns a reference to the backend that was used in this builder.

pub fn select_chain(&self) -> Option<&TSc>

Returns a reference to the select-chain that was stored in this builder.

pub fn with_opt_select_chain<USc>(
self,
select_chain_builder: impl FnOnce(&Configuration<TCfg, TGen, TCSExt>, &Arc<Backend>) -> Result<Option<USc>, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, USc, TImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>, Error>

Defines which head-of-chain strategy to use.

pub fn with_select_chain<USc>(
self,
builder: impl FnOnce(&Configuration<TCfg, TGen, TCSExt>, &Arc<Backend>) -> Result<USc, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, USc, TImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>, Error>

Defines which head-of-chain strategy to use.

pub fn with_import_queue<UImpQu>(
self,
builder: impl FnOnce(&Configuration<TCfg, TGen, TCSExt>, Arc<TCl>, Option<TSc>, Arc<TExPool>) -> Result<UImpQu, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, UImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>, Error> where
TSc: Clone,

Defines which import queue to use.

pub fn with_network_protocol<UNetP>(
self,
network_protocol_builder: impl FnOnce(&Configuration<TCfg, TGen, TCSExt>) -> Result<UNetP, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, UNetP, TExPool, TRpc, TRpcB, Backend>, Error>

Defines which network specialization protocol to use.

pub fn with_opt_finality_proof_provider(
self,
builder: impl FnOnce(Arc<TCl>, Arc<Backend>) -> Result<Option<Arc<dyn FinalityProofProvider<TBl>>>, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, Arc<dyn FinalityProofProvider<TBl>>, TNetP, TExPool, TRpc, TRpcB, Backend>, Error>

Defines which strategy to use for providing finality proofs.

pub fn with_finality_proof_provider(
self,
build: impl FnOnce(Arc<TCl>, Arc<Backend>) -> Result<Arc<dyn FinalityProofProvider<TBl>>, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, Arc<dyn FinalityProofProvider<TBl>>, TNetP, TExPool, TRpc, TRpcB, Backend>, Error>

Defines which strategy to use for providing finality proofs.

pub fn with_import_queue_and_opt_fprb<UImpQu, UFprb>(
self,
builder: impl FnOnce(&Configuration<TCfg, TGen, TCSExt>, Arc<TCl>, Arc<Backend>, Option<TFchr>, Option<TSc>, Arc<TExPool>) -> Result<(UImpQu, Option<UFprb>), Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, UImpQu, UFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>, Error> where
TSc: Clone,
TFchr: Clone,

Defines which import queue to use.

pub fn with_import_queue_and_fprb<UImpQu, UFprb>(
self,
builder: impl FnOnce(&Configuration<TCfg, TGen, TCSExt>, Arc<TCl>, Arc<Backend>, Option<TFchr>, Option<TSc>, Arc<TExPool>) -> Result<(UImpQu, UFprb), Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, UImpQu, UFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>, Error> where
TSc: Clone,
TFchr: Clone,

Defines which import queue to use.

pub fn with_transaction_pool<UExPool>(
self,
transaction_pool_builder: impl FnOnce(Options, Arc<TCl>) -> Result<UExPool, Error>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, TNetP, UExPool, TRpc, TRpcB, Backend>, Error>

Defines which transaction pool to use.

pub fn with_rpc_extensions<URpc>(
self,
rpc_ext_builder: impl FnOnce(Arc<TCl>, Arc<TExPool>) -> URpc
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, TNetP, TExPool, URpc, TRpcB, Backend>, Error>

Defines the RPC extensions to use.

pub fn with_dht_event_tx(
self,
dht_event_tx: Sender<DhtEvent>
) -> Result<ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, TCl, TFchr, TSc, TImpQu, TFprb, TFpp, TNetP, TExPool, TRpc, TRpcB, Backend>, Error>

Adds a dht event sender to builder to be used by the network to send dht events to the authority discovery module.

impl<TBl, TRtApi, TCfg, TGen, TCSExt, TBackend, TExec, TSc, TImpQu, TNetP, TExPoolApi, TRpc, TRpcB> ServiceBuilder<TBl, TRtApi, TCfg, TGen, TCSExt, Client<TBackend, TExec, TBl, TRtApi>, Arc<OnDemand<TBl>>, TSc, TImpQu, BoxFinalityProofRequestBuilder<TBl>, Arc<dyn FinalityProofProvider<TBl>>, TNetP, TransactionPool<TExPoolApi>, TRpc, TRpcB, TBackend> where
Client<TBackend, TExec, TBl, TRtApi>: ProvideRuntimeApi,
<Client<TBackend, TExec, TBl, TRtApi> as ProvideRuntimeApi>::Api: Metadata<TBl> + OffchainWorkerApi<TBl> + TaggedTransactionQueue<TBl> + SessionKeys<TBl>,
TBl: BlockT<Hash = <Blake2Hasher as Hasher>::Out>,
TRtApi: 'static + Send + Sync,
TCfg: Default,
TGen: RuntimeGenesis,
TCSExt: Extension,
TBackend: 'static + Backend<TBl, Blake2Hasher> + Send,
TExec: 'static + CallExecutor<TBl, Blake2Hasher> + Send + Sync + Clone,
TSc: Clone,
TImpQu: 'static + ImportQueue<TBl>,
TNetP: NetworkSpecialization<TBl>,
TExPoolApi: 'static + ChainApi<Block = TBl, Hash = <TBl as BlockT>::Hash>,
TRpc: RpcExtension<Metadata> + Clone,
TRpcB: RpcBuilder<TBl, TBackend, TExec, TRtApi>,

pub fn build(
self
) -> Result<NewService<TBl, Client<TBackend, TExec, TBl, TRtApi>, TSc, NetworkStatus<TBl>, NetworkService<TBl, TNetP, <TBl as BlockT>::Hash>, TransactionPool<TExPoolApi>, OffchainWorkers<Client<TBackend, TExec, TBl, TRtApi>, TBackend::OffchainStorage, TBl>>, Error>

Builds the service.