Trait substrate_service::Components

pub trait Components: Sized + 'static {
    type Factory: ServiceFactory;
    type Backend: 'static + Backend<FactoryBlock<Self::Factory>, Blake2Hasher>;
    type Executor: 'static + CallExecutor<FactoryBlock<Self::Factory>, Blake2Hasher> + Send + Sync + Clone;
    type RuntimeApi: Send + Sync;
    type RuntimeServices: ServiceTrait<Self>;
    type TransactionPoolApi: 'static + ChainApi<Hash = <FactoryBlock<Self::Factory> as BlockT>::Hash, Block = FactoryBlock<Self::Factory>>;
    type ImportQueue: ImportQueue<FactoryBlock<Self::Factory>> + 'static;
    type SelectChain: SelectChain<FactoryBlock<Self::Factory>>;
    fn build_client(
        config: &FactoryFullConfiguration<Self::Factory>, 
        executor: CodeExecutor<Self::Factory>, 
        keystore: Option<BareCryptoStorePtr>
    ) -> Result<(Arc<ComponentClient<Self>>, Option<Arc<OnDemand<FactoryBlock<Self::Factory>>>>), Error>;

    fn build_transaction_pool(
        config: TransactionPoolOptions, 
        client: Arc<ComponentClient<Self>>
    ) -> Result<TransactionPool<Self::TransactionPoolApi>, Error>;

    fn build_import_queue(
        config: &mut FactoryFullConfiguration<Self::Factory>, 
        client: Arc<ComponentClient<Self>>, 
        select_chain: Option<Self::SelectChain>
    ) -> Result<(Self::ImportQueue, Option<BoxFinalityProofRequestBuilder<FactoryBlock<Self::Factory>>>), Error>;

    fn build_finality_proof_provider(
        client: Arc<ComponentClient<Self>>
    ) -> Result<Option<Arc<dyn FinalityProofProvider<<Self::Factory as ServiceFactory>::Block>>>, Error>;

    fn build_select_chain(
        config: &mut FactoryFullConfiguration<Self::Factory>, 
        client: Arc<ComponentClient<Self>>
    ) -> Result<Option<Self::SelectChain>, Error>;
}

A collection of types and function to generalize over full / light client type.

Associated Types

type Factory: ServiceFactory

Associated service factory.

type Backend: 'static + Backend<FactoryBlock<Self::Factory>, Blake2Hasher>

Client backend.

type Executor: 'static + CallExecutor<FactoryBlock<Self::Factory>, Blake2Hasher> + Send + Sync + Clone

Client executor.

type RuntimeApi: Send + Sync

The type that implements the runtime API.

type RuntimeServices: ServiceTrait<Self>

A type that can start all runtime-dependent services.

type TransactionPoolApi: 'static + ChainApi<Hash = <FactoryBlock<Self::Factory> as BlockT>::Hash, Block = FactoryBlock<Self::Factory>>

Extrinsic pool type.

type ImportQueue: ImportQueue<FactoryBlock<Self::Factory>> + 'static

Our Import Queue

type SelectChain: SelectChain<FactoryBlock<Self::Factory>>

The Fork Choice Strategy for the chain

Loading content...
Required methods

fn build_client(
config: &FactoryFullConfiguration<Self::Factory>,
executor: CodeExecutor<Self::Factory>,
keystore: Option<BareCryptoStorePtr>
) -> Result<(Arc<ComponentClient<Self>>, Option<Arc<OnDemand<FactoryBlock<Self::Factory>>>>), Error>

Create client.

fn build_transaction_pool(
config: TransactionPoolOptions,
client: Arc<ComponentClient<Self>>
) -> Result<TransactionPool<Self::TransactionPoolApi>, Error>

Create extrinsic pool.

fn build_import_queue(
config: &mut FactoryFullConfiguration<Self::Factory>,
client: Arc<ComponentClient<Self>>,
select_chain: Option<Self::SelectChain>
) -> Result<(Self::ImportQueue, Option<BoxFinalityProofRequestBuilder<FactoryBlock<Self::Factory>>>), Error>

Build the queue that imports blocks from the network, and optionally a way for the network to build requests for proofs of finality.

fn build_finality_proof_provider(
client: Arc<ComponentClient<Self>>
) -> Result<Option<Arc<dyn FinalityProofProvider<<Self::Factory as ServiceFactory>::Block>>>, Error>

Finality proof provider for serving network requests.

fn build_select_chain(
config: &mut FactoryFullConfiguration<Self::Factory>,
client: Arc<ComponentClient<Self>>
) -> Result<Option<Self::SelectChain>, Error>

Build fork choice selector

Loading content...